Refactor: Create generic AlternativeModelPanel component to reduce code duplication
